refactor(upgrade): adds target upgrade version and refactor script (#2370)

+++ k8s/upgrades/0.8.0-0.8.1/cstor_volume_upgrade.sh
@@ -5,6 +5,8 @@
 #                                                              #
 # NOTES: Obtain the pv to upgrade via "kubectl get pv"         #
 ################################################################
+target_upgrade_version="0.8.1"
+current_version="0.8.0"
 
 function usage() {
     echo 
@@ -22,10 +24,11 @@ function setDeploymentRecreateStrategy() {
     dns=$1 # deployment namespace
     dn=$2  # deployment name
     currStrategy=`kubectl get deploy -n $dns $dn -o jsonpath="{.spec.strategy.type}"`
+    rc=$?; if [ $rc -ne 0 ]; then echo "Failed to get the deployment stratergy for $dn | Exit code: $rc"; exit; fi
 
-    if [ $currStrategy = "RollingUpdate" ]; then
+    if [ $currStrategy != "Recreate" ]; then
        kubectl patch deployment --namespace $dns --type json $dn -p "$(cat patch-strategy-recreate.json)"
-       rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
+       rc=$?; if [ $rc -ne 0 ]; then echo "Failed to patch the deployment $dn | Exit code: $rc"; exit; fi
        echo "Deployment upgrade strategy set as recreate"
     else
        echo "Deployment upgrade strategy was already set as recreate"
@@ -51,7 +54,6 @@ if [ $cas_type != "cstor" ]; then
     echo "Cstor volume not found";exit 1;
 fi
 
-pvc=`kubectl get pv $pv -o jsonpath="{.spec.claimRef.name}"`
 sc_ns=`kubectl get pv $pv -o jsonpath="{.spec.claimRef.namespace}"`
 sc_name=`kubectl get pv $pv -o jsonpath="{.spec.storageClassName}"`
 sc_res_ver=`kubectl get sc $sc_name -n $sc_ns -o jsonpath="{.metadata.resourceVersion}"`
@@ -64,29 +66,62 @@ sc_res_ver=`kubectl get sc $sc_name -n $sc_ns -o jsonpath="{.metadata.resourceVe
 # c-dep: pvc-cec8e86d-0bcc-11e8-be1c-000c298ff5fc-target        # 
 #################################################################
 
-c_dep=$(echo $pv-target)
-c_svc=$(echo $pv)
-c_vol=$(echo $pv)
-
-# Check if openebs resources exist and provisioned version is 0.8
-
-kubectl get deployment $c_dep -n $ns &>/dev/null
-rc=$?; if [ $rc -ne 0 ]; then echo "Target deployment not found: $rc"; exit; fi
-
-openebs_version=`kubectl get deployment $c_dep -n $ns -o jsonpath='{.metadata.labels.openebs\.io/version}'`
-if [ $openebs_version != "0.8.0" ]; then
-    echo "Current volumes version is not 0.8.0";exit 1;    
-fi
-
-kubectl get svc $c_svc -n $ns &>/dev/null
-rc=$?; if [ $rc -ne 0 ]; then echo "Service not found: $rc"; exit; fi
-
+c_dep=$(kubectl get deploy -n $ns -l openebs.io/persistent-volume=$pv,openebs.io/target=cstor-target -o jsonpath="{.items[*].metadata.name}")
+c_svc=$(kubectl get svc -n $ns -l openebs.io/persistent-volume=$pv,openebs.io/target-service=cstor-target-svc -o jsonpath="{.items[*].metadata.name}")
+c_vol=$(kubectl get cstorvolumes -l openebs.io/persistent-volume=$pv -n $ns -o jsonpath="{.items[*].metadata.name}")
+c_replicas=$(kubectl get cvr -n openebs -l openebs.io/persistent-volume=$pv -o jsonpath="{range .items[*]}{@.metadata.name};{end}" | tr ";" "\n")
 
 # Fetch the older target and replica - ReplicaSet objects which need to be 
 # deleted before upgrading. If not deleted, the new pods will be stuck in 
 # creating state - due to affinity rules. 
 
-c_rs=$(kubectl get rs -o name --namespace $ns | grep $c_dep | cut -d '/' -f 2)
+c_rs=$(kubectl get rs -n $ns -o name -l openebs.io/persistent-volume=$pv | cut -d '/' -f 2)
+
+
+# Check if openebs resources exist and provisioned version is 0.8
+
+if [[ -z $c_rs ]]; then
+    echo "Target Replica set not found"; exit 1;
+fi
+
+if [[ -z $c_dep ]]; then 
+    echo "Target deployment not found"; exit 1;
+fi
+
+if [[ -z $c_svc ]]; then
+    echo "Target svc not found";exit 1;
+fi
+
+if [[ -z $c_vol ]]; then
+    echo "CstorVolumes CR not found"; exit 1;
+fi
+
+if [[ -z $c_replicas ]]; then
+    echo "Cstor Volume Replica CR not found"; exit 1;
+fi
+
+controller_version=`kubectl get deployment $c_dep -n $ns -o jsonpath='{.metadata.labels.openebs\.io/version}'`
+if [[ "$controller_version" != "$current_version" ]] && [[ "$controller_version" != "$target_upgrade_version" ]] ; then
+    echo "Current cstor target deloyment $c_dep version is not $current_version or $target_upgrade_version";exit 1;    
+fi
+
+controller_service_version=`kubectl get svc $c_svc -n $ns -o jsonpath='{.metadata.labels.openebs\.io/version}'`
+if [[ "$controller_service_version" != "$current_version" ]] && [[ "$controller_service_version" != "$target_upgrade_version" ]]; then
+    echo "Current cstor target service $c_svc version is not $current_version or $target_upgrade_version";exit 1;    
+fi
+
+cstor_volume_version=`kubectl get cstorvolumes $c_vol -n $ns -o jsonpath='{.metadata.labels.openebs\.io/version}'`
+if [[ "$cstor_volume_version" != "$current_version" ]] && [[ "$cstor_volume_version" != "$target_upgrade_version" ]]; then
+    echo "Current cstor volume  $c_vol version is not $current_version or $target_upgrade_version";exit 1;    
+fi
+
+for replica in $c_replicas
+do
+    replica_version=`kubectl get cvr $replica -n openebs -o jsonpath='{.metadata.labels.openebs\.io/version}'`
+    if [[ "$replica_version" != "$current_version" ]] && [[ "$replica_version" != "$target_upgrade_version" ]]; then
+        echo "CStor volume replica $replica version is not $current_version"; exit 1;
+    fi
+done
 
 
 ################################################################ 
@@ -97,10 +132,10 @@ c_rs=$(kubectl get rs -o name --namespace $ns | grep $c_dep | cut -d '/' -f 2)
 # previous step                                                #  
 ################################################################
 
-sed "s/@sc_name/$sc_name/g" cstor-target-patch.tpl.json | sed -u "s/@sc_resource_version/$sc_res_ver/g" > cstor-target-patch.json
-sed "s/@sc_name/$sc_name/g" cstor-target-svc-patch.tpl.json | sed -u "s/@sc_resource_version/$sc_res_ver/g" > cstor-target-svc-patch.json
-sed "s/@sc_name/$sc_name/g" cstor-volume-patch.tpl.json | sed -u "s/@sc_resource_version/$sc_res_ver/g" > cstor-volume-patch.json
-sed "s/@sc_name/$sc_name/g" cstor-volume-replica-patch.tpl.json | sed -u "s/@sc_resource_version/$sc_res_ver/g" > cstor-volume-replica-patch.json
+sed "s/@sc_name@/$sc_name/g" cstor-target-patch.tpl.json | sed -u "s/@sc_resource_version@/$sc_res_ver/g" | sed -u "s/@target_version@/$target_upgrade_version/g" > cstor-target-patch.json
+sed "s/@sc_name@/$sc_name/g" cstor-target-svc-patch.tpl.json | sed -u "s/@sc_resource_version@/$sc_res_ver/g" | sed -u "s/@target_version@/$target_upgrade_version/g" > cstor-target-svc-patch.json
+sed "s/@sc_name@/$sc_name/g" cstor-volume-patch.tpl.json | sed -u "s/@sc_resource_version@/$sc_res_ver/g" | sed -u "s/@target_version@/$target_upgrade_version/g" > cstor-volume-patch.json
+sed "s/@sc_name@/$sc_name/g" cstor-volume-replica-patch.tpl.json | sed -u "s/@sc_resource_version@/$sc_res_ver/g" | sed -u "s/@target_version@/$target_upgrade_version/g" > cstor-volume-replica-patch.json
 
 #################################################################################
 # STEP: Patch OpenEBS volume deployments (cstor-target, cstor-svc)              #  
@@ -108,46 +143,56 @@ sed "s/@sc_name/$sc_name/g" cstor-volume-replica-patch.tpl.json | sed -u "s/@sc_
 
 
 # #### PATCH TARGET DEPLOYMENT ####
-echo "Upgrading Target Deployment to 0.8.1"
 
-# Setting deployment startegy to recreate
-setDeploymentRecreateStrategy $ns $c_dep
+if [[ "$controller_version" != "$target_upgrade_version" ]]; then
+    echo "Upgrading Target Deployment to $target_upgrade_version"
 
-kubectl patch deployment  --namespace $ns $c_dep -p "$(cat cstor-target-patch.json)" 
-rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
+    # Setting deployment startegy to recreate
+    setDeploymentRecreateStrategy $ns $c_dep
 
+    kubectl patch deployment  --namespace $ns $c_dep -p "$(cat cstor-target-patch.json)" 
+    rc=$?; if [ $rc -ne 0 ]; then echo "Failed to patch cstor target deployment $c_dep | Exit code: $rc"; exit; fi
 
+    kubectl delete rs $c_rs --namespace $ns
+    rc=$?; if [ $rc -ne 0 ]; then echo "Failed to delete cstor replica set $c_rs | Exit code: $rc"; exit; fi
 
-kubectl delete rs $c_rs --namespace $ns
-rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
-
-
-rollout_status=$(kubectl rollout status --namespace $ns  deployment/$c_dep)
-rc=$?; if [[ ($rc -ne 0) || !($rollout_status =~ "successfully rolled out") ]];
-then echo "Upgrade failed | ERROR: $rc"; exit; fi
-
-
+    rollout_status=$(kubectl rollout status --namespace $ns  deployment/$c_dep)
+    rc=$?; if [[ ($rc -ne 0) || !($rollout_status =~ "successfully rolled out") ]];
+    then echo "Failed to rollout for deployment $c_dep | Exit code: $rc"; exit; fi
+else
+    echo "Target deployment $c_dep is already at $target_upgrade_version"
+fi
 
 # #### PATCH TARGET SERVICE ####
-echo "Upgrading Target Service to 0.8.1"
-kubectl patch service --namespace $ns $c_svc -p "$(cat cstor-target-svc-patch.json)"
-rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
+if [[ "$controller_service_version" != "$target_upgrade_version" ]]; then
+    echo "Upgrading Target Service to $target_upgrade_version"
+    kubectl patch service --namespace $ns $c_svc -p "$(cat cstor-target-svc-patch.json)"
+    rc=$?; if [ $rc -ne 0 ]; then echo "Failed to patch service $svc | Exit code: $rc"; exit; fi
+else 
+    echo "Target service $c_svc is already at $target_upgrade_version"
+fi
 
 # #### PATCH CSTOR Volume CR ####
-echo "Upgrading cstor volume CR to 0.8.1"
-kubectl patch cstorvolume --namespace $ns $c_svc -p "$(cat cstor-volume-patch.json)" --type=merge
-rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
+if [[ "$cstor_volume_version" != "$target_upgrade_version" ]]; then
+    echo "Upgrading cstor volume CR to $target_upgrade_version"
+    kubectl patch cstorvolume --namespace $ns $c_vol -p "$(cat cstor-volume-patch.json)" --type=merge
+    rc=$?; if [ $rc -ne 0 ]; then echo "Failed to patch cstor volumes CR $c_vol | Exit code: $rc"; exit; fi
+else
+    echo "CStor volume CR  $c_vol is already at $target_upgrade_version"
+fi
 
 # #### PATCH CSTOR Volume Replica CR ####
-echo "Upgrading cstor volume replicas CR to 0.8.1"
-replicas=$(kubectl get cvr -n $ns -l cstorvolume.openebs.io/name=$pv -o jsonpath="{range .items[*]}{@.metadata.name};{end}" | tr ";" "\n")
 
-for replica in $replicas
+for replica in $c_replicas
 do
-    echo "Patching replic: $replica"
-    kubectl patch cvr $replica --namespace openebs -p "$(cat cstor-volume-replica-patch.json)" --type=merge
-    rc=$?; if [ $rc -ne 0 ]; then echo "Upgrade failed | ERROR: $rc"; exit; fi
-    echo "Successfully updated replica: $replica"
+    if [[ "`kubectl get cvr $replica -n openebs -o jsonpath='{.metadata.labels.openebs\.io/version}'`" != "$target_upgrade_version" ]]; then
+        echo "Upgrading cstor volume replica $replica to $target_upgrade_version"
+        kubectl patch cvr $replica --namespace openebs -p "$(cat cstor-volume-replica-patch.json)" --type=merge
+        rc=$?; if [ $rc -ne 0 ]; then echo "Failed to patch CstorVolumeReplica $replica | Exit code: $rc"; exit; fi
+        echo "Successfully updated replica: $replica"
+    else
+        echo "cstor replica  $replica is already at $target_upgrade_version"
+    fi
 done
 
 echo "Clearing temporary files"
@@ -156,6 +201,6 @@ rm cstor-target-svc-patch.json
 rm cstor-volume-patch.json
 rm cstor-volume-replica-patch.json
 
-echo "Successfully upgraded $pv to 0.8.1 Please run your application checks."
+echo "Successfully upgraded $pv to $target_upgrade_version Please run your application checks."
 exit 0
